+++ src/main/java/adega/de/vinhos/adegadevinhos/dto/AdegaDTO.java
@@ -1,0 +1,13 @@
+package adega.de.vinhos.adegadevinhos.dto;
+
+import adega.de.vinhos.adegadevinhos.util.TranslationConstants;
+import jakarta.validation.constraints.NotEmpty;
+import lombok.Data;
+
+@Data
+public class AdegaDTO {
+    private Long id;
+    @NotEmpty(message = TranslationConstants.NOME_ADEGA_NOT_NULL)
+    private String nome;
+    private Integer capacidade;
+}