+++ src/main/java/adega/de/vinhos/adegadevinhos/util/TranslationConstants.java
@@ -8,4 +8,6 @@ public class TranslationConstants {
     public static final String BAD_REQUEST_INVALID_FIELDS = "Bad Request Exception, os campos não podem ser nulos";
     public static final String CHECK_FIELDS_ERROR = "Verifique o preenchimento do(s) campo(s)";
 
+    public static final String NOME_ADEGA_NOT_NULL = "O nome da adega não pode ser nulo";
+    public static final String ADEGA_NAO_ENCONTRADA = "Adega não encontrada";
 }