+++ src/test/java/adega/de/vinhos/adegadevinhos/AdegaDeVinhosApplicationTests.java
@@ -1,9 +1,8 @@
 package adega.de.vinhos.adegadevinhos;
 
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
+//@SpringBootTest
 class AdegaDeVinhosApplicationTests {
 
 	@Test